test: finalize lifecycle ownership cleanup (#170)
